moved tests to reth-trie-db
