--- crates/trie/tests/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "reth-trie-tests"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-exclude.workspace = true
-
-[dev-dependencies]
-reth-db = { workspace = true, features = ["test-utils", "mdbx"] }
-reth-primitives.workspace = true
-reth-storage-errors.workspace = true
-reth-trie.workspace = true
-reth-trie-db.workspace = true
-
-alloy-primitives.workspace = true
-alloy-trie.workspace = true
-
-nybbles.workspace = true
-
-[lints]
-workspace = true
